requirements.txt
mss>=9.0.1
opencv-python>=4.9.0
numpy>=1.26.0
anthropic>=0.40.0
pyttsx3>=2.90
PyYAML>=6.0.1
Pillow>=10.0.0
customtkinter>=5.2.0
python-dotenv>=1.0.0
scipy>=1.12.0
scikit-learn>=1.4.0
soundcard>=0.4.2
pynput>=1.7.6
